fetch-system-image: add a synthentic ni-arch.conf
Real targets end up running ni-arch-gen on-target as part of their image
install. We have access to that (it's in the feed) but it's a bit of a
pain to deal with (requires python3, and python3 isn't in base images--
something that the image install process deals with by manually installing
a whole slew of python ipks from outside the feed, ick).

In theory we should have been fine just with the core2-64 and x64 arches,
except some NI IPKs went hard on using target-specific definitions, such
as with `ni-daqmx`, so those packages end up being invisible unless we add
an "x64-PXI" arch or similar. :(